fix: fix avatar uploader test and create homePage.cy.ts file
